api: Adjust `MessageSnapshot` type to align with docs; add jsdoc.
The array of these that this endpoint returns collectively make up the
history, but each one isn't a history.  The doc calls them "snapshot
objects", so follow its naming.

Also the doc indicates that three more of these properties can be
missing.  It isn't super clear about that, unfortunately, but flag it
so it won't be a surprise.

And this looks a lot like the `MessageEdit` object, which serves a
similar purpose; cross-reference them to help avoid confusion.