notif: Store device token globally and ephemerally.
This is the appropriate place to keep the answer to the question of
what our device token *really* is.  It's independent of the Zulip app
server, so no need to wait until the server acks it or to invalidate
it on switching accounts; and it may change between runs, so it'd be a
bug to get it from our own persistent store.