notif: Store server-acked token in `account`, rather than `realm`.
The main effect of this is that we remember this information across
switching accounts.  The server doesn't forget it just because we
switched accounts -- it might because we ask it to, but that's not the
same thing at all and we already model it separately with the UNACK
action type.  Because the point of this field is to model what the
server knows, that means we shouldn't forget it either.

Also give it a more specific name to distinguish its semantics from
`session.pushToken`.