notif: Don't stop notifications on switching accounts.
The experience when such a notification arrives is pretty broken,
yes... but I think it's less broken than having it not arrive at all.

In particular, if you see or hear the notification while at your desk
and go read the message on the desktop, you'll never experience the
brokenness at all.

This also simplifies the state machine governing `ackedPushToken` on
an account: now the only time it turns back to `null` is when the user
logs out.