android notif: Identify more directly the system default sound.
When `DEFAULT_SOUND` (or `DEFAULT_ALL`) is set on the notification,
the place the system actually looks to find the default value is
here, not RingtoneManager.  See e.g. (search for DEFAULT_SOUND)
  https://android.googlesource.com/platform/frameworks/base/+/pie-release/services/core/java/com/android/server/notification/NotificationRecord.java

or for the logic that actually runs in earlier Android versions:
  https://android.googlesource.com/platform/frameworks/base/+/nougat-release/services/core/java/com/android/server/notification/NotificationManagerService.java

Inspection of the RingtoneManager source finds that it also returns
this same value; but no need to involve the extra indirection and
have to wonder if that's definitely always true.