store: Buffer thunk actions, as well as plain objects.
Our action functions are rarely, if ever, written for the possibility
that they might be run before the Redux state has been rehydrated from
persistent storage.

For the vast majority of them this is impossible in any case, because
they're invoked from code which will only run from some UI that isn't
the loading screen -- which means after our `nav` reducer responds to
a REHYDRATE action -- or in the background in response to
`needsInitialFetch` becoming true.  For the rest, this is an edge case
that only makes it more complicated to reason about the behavior -- or
more realistically, that simply doesn't get reasoned about at all.

At a quick test, starting up a logged-in install of the app, the only
action functions that got run before REHYDRATE were `reportPresence`
and `trySendMessages`.  Neither does anything useful when run then.
Both are harmless, though perhaps redundant, if delayed and run later.

If in the future we want to let specific action functions skip the
buffer, we can do that easily enough by marking them with some
property and giving `createActionBuffer` a `passthrough` option that
checks for that property.  For the reasons above it's important that
any such mechanism remain opt-in, not opt-out.