emoji: Rename all `*RealmEmoji*` to `*ImageEmoji*`.
So that all this selector can also take care of Zulip extra emojis.

We don't yet *have* any Zulip extra emoji supported in the app; so
"all image emoji" and "all realm emoji" are the same set of emoji.
We'll be adding (the one) Zulip extra emoji in an upcoming commit.