redux: Get real types on our reducers and our store.
This unfortunately doesn't connect the dots all the way to `connect`,
because the trail ends at `Provider` in the `react-redux` libdef.

It also doesn't quite connect the dots to our individual reducer
functions, because the `redux` libdef has an `any` in the type of
`combineReducers`.  There, though, it's a single-token fix to write
down the correct type...

... and when I do that, together with these changes, Flow has enough
information that it catches bogus reducer types!  Like the ones we
fixed a few commits ago.

Because the libdef comes from `flow-typed` and we have a script to
sync from that unconditionally, it'll take some extra steps to
actually apply that fix.  Not going to do that today; perhaps sometime
soon I'll look into getting the fix into `flow-typed` upstream.

Meanwhile, no reason not to merge the trail sections that do work.