actions: Clarify to Flow that these constants are type-tags.
When Flow sees a definition like `const FOO = 'FOO'`, it knows for
some purposes that `FOO` is specifically `'FOO'`... but for others, it
apparently broadens it to `string`.  This is facebook/flow#2639.

One consequence is that when we say

  switch (action.type) {
    case NOT_FOO_LOL:
      // ... use `action` ...
      break;

    // ...
  }

with a `case` statement for a value of `type` that we've claimed is
*impossible*... Flow quietly just decides that `action` has type
`empty` and there's nothing to worry about.

Then if our code goes and passes that `action` value to *any function
at all*, Flow considers it valid.  After all, there are no values of
type `empty` (that's what `empty` is all about)... so it's true that
every value of type `empty` is a valid value of whatever type it is
that the function expected.

So, that's pretty unfortunate.  Fortunately, there is a workaround
that causes Flow to behave better here: duplicate the value into the
type, like `const FOO: 'FOO' = 'FOO'`.  (This trick due to @geraldyeo
and @fagerbua, in comments on facebook/flow#2377 -- thanks!)

Do that, then.

---

Made the edit with

  perl -i -0pe \
    's/export const ([A-Z_]+)\K =/: "$1" =/g' \
    src/actionConstants.js

and then manual fixup on the one exceptional case `REHYDRATE`,
which Flow kindly pointed out.

We had a whole bunch of type errors of the kind this hides; the bulk
of them were for Event*Action types, and their fixes went into the
parent commit.  This commit adds one more missing action type which a
number of reducers had type errors on, and a few other simple fixes.

PS: A quick tip for reading this diff: `git log -p --patience`.