actions: Make EventAction complete.
This resolves the long list of type errors Flow saw in eventToAction.

It also resolves an even longer list of type errors in reducers, which
have been hidden by a curious Flow issue; the next commit applies a
workaround for that issue, which exposes all those errors and is how I
found these types were missing from the EventAction union.

As a nice little bonus, it lets us simplify the definition of Action.