event: Remove the confused `user` property from subscription actions.
This is completely meaningless for the events that aren't peer_add
and peer_remove; whatever we do for this should be specific to the
event/action types where it makes sense.

And we don't currently do anything with those events where it does
make sense -- so until we do, just cut it.