event: Conform stream action names to the server's naming.
The server API has some inconsistent naming, using `create`/`delete`
here on stream events while other parts of the API (like events on
reactions, users, and user groups) use `add`/`remove`.

We'd chosen to make the names within our code consistent between
different types of items, at the cost of making them inconsistent here
between our code and the server API.

But after recent refactors, our names related to these events have a
greatly reduced presence in our code.  The parent commit removed
action-type-constant names like EVENT_STREAM_ADD; and recent commits
made action-type names like EventStreamAddAction local to actionTypes
and not exported:
  443e2bf80 redux: Use type inference in reducers.
  9707e797b actions: Make almost all action types local

As a result, basically the only role of these names at this point is
that Flow can show them in error messages to help keep them readable.
They never need to be typed in random places, so there's no longer
much benefit to making them consistent with our other names... and on
the other hand, in a non-exhaustiveness error, it's helpful for the
name of the type to match the value that another `case` is needed for,
i.e. the `op` value in the server event.

So, rename these to conform to the server's naming.