compose box: Cut remaining two vague types.
For one of these, we don't even need an annotation!  That's how it
should be for a callback like this, and thankfully the libdef for View
has an appropriate type so it works out.

For the other, the libdef for TextInput has an `any` in the type;
so cover that over with a better type.