MultilineInput: Cut this trivial, and misleading, wrapper.
This does just two things on top of `Input`:
 * Set `multiline`.
 * Set `underlineColorAndroid` to `"transparent"`.

The first one is just as easily spelled `<Input multiline` as
`<MultilineInput`, and writing it that way is much more transparent
to read and understand.

If the second one were dealing with some gotcha about multiline inputs
and the Android/Material text-input underlines, so that this component
encoded an important lesson about how to handle multiline text inputs,
that would be one thing and could be a very good reason to have this
component.

But in fact the opposite is the case: setting no-underline is really
about the specific styling we have around the inputs in the compose
box, where the input's borders are marked by a contrast in the
background color.  We have the exact same choice on the topic input --
which is not multiline -- for the same reason.  And on the other hand
if we were to use a multiline input in a more typical configuration
where it didn't have marked borders, then this MultilineInput with its
hiding of the underline would look completely wrong.

So, move the `underlineColorAndroid="transparent"` bit to where it
makes sense, which is the caller.  And then just inline the remaining
logic, i.e. "<Input multiline ...>", and delete the component.