chat: Render `ComposeBox` only if `canSend`.
Fixes #3295.

If user can not send message in that narrow then it is better to not
render ComposeBox component at all.  This is the normal React pattern
for conditional rendering.

Since commit 824c908cc we've used instead a hack to make the component
render in React but not actually get displayed.  The motivation for
that, in turn, was that we had a hack using `componentDidMount` to
initialize the two TextInput components; and we'd been creating a
ComposeBox component but then having its `render` just return `null`,
so that hack wouldn't fire a second time if the compose box got
hidden, then unhidden again.

By putting ComposeBox itself inside the conditional render, we make it
so even that `componentDidMount` hack would work.  And meanwhile in
3ab1b7575 we recently removed that hack anyway in favor of passing
`defaultValue` to the input components.

---

Empirically, this fixes #3295.  Why is that, and how did we find
this fix?  A few observations on the issue:

* Only reproducible when compose box is shown after message loading
  (GIF like loader) after narrowing on Android.  Can't reproduce on
  narrows whose messages are cached and immediately visible after
  narrowing.

* If on the topic input (a) the `this.getCanSelectTopic()` condition
  is removed and (b) the `multiline` prop is set, it starts showing
  the issue, just like the message input.

So from this it can be inferred that:

* It is a `multiline` input issue, which occurs when its parent first
  got rendered with `display: none`.  Normally it is not reproducible
  on topic input because at the time of first render (when style is
  `display: none`), topic input is not rendered (because
  `this.getCanSelectTopic()` is false).

So to fix this issue, use the normal React pattern for conditional
rendering, so that the input component doesn't get created at all
until we want to show it.