notif: Push messy iOS-only format-munging logic inward, to cancel out.
On Android, we've taken over the entire path to generating a
notification-opened event, and so we've made the object we get handed
in that event be exactly our `Notification` object, with no extra goo.

So, no need for any of the messy unpacking the wix library bequeathed
us, nor the wrapping in `getData: () => ...` which we added in the
course of recent refactoring in order to keep this codepath's behavior
unchanged.

The iOS case is left unchanged for now.