nullObjects: Cut nullFunction.
The main way we used to use this was to pass as callbacks like
`onPress` where we didn't have anything we actually wanted to run in
the callback.

But the conventional way that React Native upstream components accept
for indicating that is to simply pass `undefined`.  We've adopted the
same convention in our many components that take a callback prop which
they pass on to some upstream component, and then it's been natural to
do the same in the components where we accept a callback we'll call
ourselves.

This `undefined` convention has the advantage of transparency -- any
JavaScript code can easily inspect the value (simply test it for
truthiness) and determine that it isn't a function that's going to
actually do anything when called, potentially saving some effort.  It
also greatly simplifies invoking an interface that takes a number of
optional callbacks -- it's much simpler and clearer to simply omit
those properties entirely, rather than mention each of them only to
pass a no-op value.

In the preceding few commits we eliminated the handful of remaining
references to this value (the majority of them in code that we already
never invoked); now remove it.