android notif: Upgrade to FCM (Firebase), from GCM.
Basically just followed the upstream instructions for this migration:
  https://developers.google.com/cloud-messaging/android/android-migrate-fcm

And tested, and it works!

In terms of those instructions:
  1. I did this step, except skipped adding the plugin; details below.
  2. Included in this commit.
  3. Included in this commit.
  4. This is server-side; will do separately.
  5. Included in this commit.  The first two of these four apply to us.
     See the further docs linked there for each of them.

The FCM API has some nice cleanups relative to GCM.  Note how
RegistrationIntentService and InstanceIdListenerService get collapsed
into one service; and also on the message-receiving side we get to
take a service out of our source manifest.

The main point of divergence from the instructions is that I didn't
want to use the Google Services Gradle Plugin and have it process a
`google-services.json`.  The plugin wants to configure a bunch of
other services too, and doesn't make it super transparent what it's
doing; given that this isn't the kind of app that's stuffed full of
Google-flavored analytics, I wanted to keep things more explicit.

So instead I put the relevant values into a new file `firebase.xml`,
following the useful parts of what the plugin would do.  Details on
those, and how to recompute them from a `google-services.json`:
  https://developers.google.com/android/guides/google-services-plugin

And a doc that adds a bit more context:
  https://firebase.google.com/docs/android/setup#manually_add_firebase