message fetch: Make some `?foo` types more specific.
In Flow the type `?foo` translates, unfortunately, to the type
`foo | void | null` -- which is almost always broader than is
actually intended.  This can make very natural (and correct) code
like `if (thing !== undefined) ...` fail to type-check because
we've given up the knowledge that it won't be `null` instead.
So it seems a good practice to avoid it and explicitly write
either `foo | void` or `foo | null` as appropriate.

Apply that to a couple of recently-added examples.

The consequences aren't a big deal here where the underlying type
is `boolean`, because we'll mainly be testing for truthiness
anyway; but it's simplest to make the practice consistent.