nav: Cut out fudge-fallbacks on getStateForRoute.
These will fail only if we pass a route which isn't actually one of
our routes -- so if one of them ever fails, it always fails.  Best to
fail hard and catch the issue at development time.

Conversely, many of these fallbacks would themselves lead to pretty
severe symptoms... like REHYDRATE leaving us at the loading screen
instead of navigating to the actual app UI.  Even if we did somehow
ship a bug that caused that fallback to fire, it's probably less
frustrating for the user to just crash.

That example also illustrates one of the classic dangers of this kind
of fudge-fallback for a condition that should be statically
impossible: because they're impossible, they tend not to get tested or
even carefully thought through.  It's unlikely anyone ever actually
thought through what we should do if navigation to `'main'` fails; if
they had, then "just keep the loading screen forever" is probably not
the answer they would have chosen.  But that's the answer that we had,
via this fallback.