tests: Remove Coveralls.
We haven't been really using this for a long time, I think.

And when attempting to reproduce a CI build locally, it causes
failure with a confusing and unhelpful error message:

  /home/greg/z/mobile/node_modules/coveralls/bin/coveralls.js:18
          throw err;
          ^
  Bad response: 422 {"message":"Couldn't find a repository matching this job.","error":true}

We could certainly find a way to work around that; but since we
aren't really using the tool anyway, just drop it instead.