android notif: Parse `remove` items robustly to malformed message IDs.
Without this, we just crash if we get something with a message ID in a
format we weren't expecting... which I was reminded of in development
when an early draft version of the server code for batching IDs into
`zulip_message_ids` sent malformed data.

This won't typically happen in real life, because deployed servers
generally don't send malformed data... but if they do, we certainly
shouldn't crash.  Instead, log a proper warning and carry on.

This is a long-pre-existing problem elsewhere in this notifications
code, which we should go back and clean up.  For now, just set an
example of how this can be done.