TitleStream: Move single-use styles into file.
Also replace the rather confusing combination of `navWrapper` with
another style that overrides just some of its alignment properties,
instead just writing the whole set of alignment properties fresh.

As a bonus, this avoids creating a new `style` value on each re-render
and causing the child View to unnecessarily re-render in turn.