types: Provide an alias for the very common `User | RealmBot`.
We have a fair number of uses of this disjunction... and we should
probably have more, as it's likely a lot of our references to `User`
should include cross-realm bots too.  Provide an alias to help make
it easier to do.

This will extra help when we give `RealmBot` a less misleading name,
because `CrossRealmBot` is longer.