android notif: Group "recipient" fields in an ADT, to crunchify more.
This makes our "crunchy shell" crunchy in an additional area where up
to this point we'd been soft (and prone to crashing, or other buggy
behavior, on surprises): now the "stream" and "topic" fields are
required for a stream message.

Perhaps of more importance in this case: it makes the interrelationships
of all these fields explicit, making the application code that refers
to them easier to reason about.