android notif: Kotlinize and crunchify parsing `remove` messages.
The two methods `parseMessageIds` and `parseMessageIdInto`
were auto-converted by Android Studio.  I had it convert
FCMPushNotifications.java, copied out these two methods,
and discarded the rest of the conversion.

This change also brings the parsing into the same design as
we have for MessageFcmMessage.  The Kotlin conversion is
verbatim except for two adaptations I made for that:
 * turned the `Log.w` into a `throw`
 * added a direct `null` check on `zulip_message_id`.
   (Previously if it was missing we'd let Integer.parseInt
   throw an exception, and catch that.  It's important that
   it's OK for that field to be missing and only the plural
   `zulip_message_ids` present.)

Note the wacky use of dropLastWhile!  Empirically, Java's
String#split seems to really behave that way; yikes.  So
that's part of a correct, faithful conversion; good work,
IntelliJ, in getting that detail right.

In the next commit, we'll turn this into idiomatic Kotlin,
which will be a lot shorter as well as cleaning up that
semantic edge case.