android notif: Put React behind our "crunchy shell".
We now have a nice crunchy shell that provides a clean version
of the notification data, in a known and reasonable format.
It enables application code to take shelter, in a "soft center",
from the complexity of all the arbitrary stuff we might receive
from the server.

Our Android-native code lives inside that soft center, but
we're still sending React the raw messiness of the data from
the network.  Send it some nice parsed data instead, moving it
into the "soft center".

This was the last use of the `bundle` member, so cut it out.
Now we don't even keep around the raw unparsed data.