android notif: Skip `Bundle` in parsing FCM messages.
FCM gives us the message's data as a `Map<String, String>`.  We've
been converting that to an Android `Bundle`, and handing that to our
parsing code.  There's no real need for a `Bundle` here; its use is
basically a legacy of the fact that GCM gave us a `Bundle`, and (the
code that turned into) our parsing code was written for that.

Leave it as a `Map`, and convert the code that consumes it.  This is
a boring near-mechanical change, and a pure refactor.