ZulipButton: Drop redundant `style` passed to Touchable.
The `flex` property has no effect here, because this is the only
child of a parent View with a fixed height.

The `alignSelf` property doesn't seem to have any effect either --
empirically we get the same result without it, using either kind of
touchable.

The `justifyContent` and `alignItems` properties are redundant with
the same properties which we pass to the child.