webview js: Fix unexpected `click` on long-press on iOS.
The long press logic is very subtle, as described in `js.js`, and
the logic around `lastTouchEventTimestamp` is even more obscure.
There was a bug in iOS, where the a shortPress, or a `click`, was
triggered after releasing the finger after a `longPress` event.

This commit fixes this bug by introducing a flag which detects
whether a long press was triggered before the detection of this
click.  While this could be potentially implemented without an extra
flag and with the already existing `lastTouchEventTimestamp`, on
trying to implement it, there arise a slew of other bugs.  Besides,
if a new flag was not used, the logic around
`lastTouchEventTimestamp` would become unreadable.

Fixes #3429.