api: Fix uses of `Object` that really mean `{}`.
What all of these signatures are really trying to say is that they
require the parameter in question to be an object (in one case, unless
it's null or undefined)... and that they make no further assumptions
about what might be in it.

That type is spelled `{}`.  Every object is a value of type `{}`.
The type `Object` is nearly the opposite -- it allows the consumer
to make *any* assumption it likes, rather than permitting *no*
assumptions.

So, write `{}`.

One line requires a fixme because of a somewhat awkward Flow behavior:
we're scrupulously checking whether a certain property is present and
has a certain type, but even the check produces a type error.