recipient: Cut unreachable nonsense `outbox` case.
Flow 0.92 points out that our types say this is unreachable:

  All branches are incompatible:
   • Either string [1] is incompatible with enum [2].
   • Or string [1] is incompatible with enum [3].

where the references show that both Message and Outbox have

    type: 'stream' | 'private',

Moreover, if this case *were* reachable, what it's saying is total
nonsense.  Two messages have the same recipient if they happen to both
be outbox messages?!

So, thanks Flow!  And cut it out.