status bar: Write a more specific type for `orientation`.
Another error found by experimenting with a better react-redux libdef
that allows the type-checker to connect the dots across `connect`
calls.

In this case, the correct type is a subtype of what we had, so we
could also have written something like `+orientation: string` to just
clarify it was an input-only type.  But `Orientation` better describes
the actual assumptions this code is written for.