Screen: Fix some prop types to allow subtypes.
Really the most natural thing here would be to mark the entire `Props`
type as read-only... except that I'm not yet sure that interacts
correctly with `connect` for the props that that interacts with.  Only
a handful of props have types with useful subtypes anyway, so we can
just mark those individually.

This fixes errors that appear when experimenting with a libdef that
makes `connect` calls actually type-checkable.