stream list: Fix erroneous type on `streams` prop.
The real type here is pretty sadly complex -- but far better to have
that explicit in the code than to have a type annotation that lies to
us and says it's simpler.

This is a downstream discovery from the fix in the parent commit, and
the experimental improved `connect` libdef which surfaced that error.