react-redux: Add `dispatch` prop to all components that receive it.
After all, we've written exact object types here -- so it's a
type error to end up passing a props object that has additional,
unmentioned properties like `dispatch`.

That class of type error is one that has gone un-complained-about
to date because we've lacked much of a useful libdef for `connect`;
but experimenting with the now-available better one generates a
bunch of error messages about it.  So, fix it.

Done almost entirely by machine:

  $ git grep -l 'connect.*react-redux' src/ \
    | xargs -r git grep -L dispatch: \
    | xargs -r perl -i -0pe '
        s/^(?:export )?type Props = .*\{\|?\K/ dispatch: Dispatch, /m;
        s/,?\s*\} from \S*types/, Dispatch $&/ms
      ' \
    && tools/fmt

Then manually fixed up one file, LegalScreen.js, to add the needed
Dispatch import.  (It was the one file that didn't already import
something else from `types`, which the command above otherwise
piggybacked on.)