RealmScreen: Clean up type error with undefined vs. empty-string.
Found by Flow on marking the file strict-local: if `error` were to be
the empty string, we'd end up trying to use that as a React child,
which would not go well.

Also use `null` rather than `undefined` for the "no error" state;
we really do mean "there is no error", not "no data on whether
there's an error", and so it seems a bit cleaner and more idiomatic.
(To the extent there is a clean pattern for distinguishing the two
in JavaScript.)