MessageList: Propagate awareness of outbox case through types.
Lots of code around the message list has been confused about the
type of `messages` and related values -- the truth is that it can
contain `Outbox` objects as well as `Message`, and it's even
sprinkled with conditionals checking for that, but the types have
been labeled just `Message`.

These errors have up to now been hidden by the lack of types on
`connectActionSheet`, which defeats type-checking at the
selectors invoked by MessageList.  With our spiffy new well-typed
`connect`, and an upcoming `SelectorProps` annotation to bypass
`connectActionSheet`, we can see these errors -- so, fix them.

This commit just revises a bunch of type annotations to
acknowledge the type as `Message | Outbox`.  In the parent
commit, we made an actual runtime logic change to fix a potential
live issue this exposed.