android notif: Collapse the trivial `fetchAvatar` with `fetch`.
Rather than have one function throw `IOException`, and a trivial
wrapper convert that to returning `null`, just present the API we
prefer here in the first place: return `null`.

Also, give the function a more accurate name.  It's much more specific
than generically "fetching" just anything; but it doesn't know
anything in particular about avatars.  What it does know is that it's
going to turn the response it gets into a bitmap -- so `fetchBitmap`.