SubscriptionsCard: Clean up and simplify props.
Remove several lines of code that have had no effect for a very
long time, since we stopped having a navigation drawer.

The purpose of the code being removed was to show a subscription as
'selected' if it was the currently active narrow (this component was
rendered in the drawer.)

This component now only serves as a Tab in StreamTabs. It never
receives a value for the 'narrow' property, thus it remains the
default `[]`. This in turn always sets `selected` to `false`.

None of the removed code had any effect except cluttering the code.