flow: Move `ConnectionInfo` type to narrow scope, and match name.
The linked docs page uses the name `ConnectionType` specifically for
the enum found in the `type` property of this type.  The other
property's type is `EffectiveConnectionType`.  The object type
containing both isn't given a name, but parameters of that type are
called `connectionInfo`, which seems apt.

So, avoid reusing the upstream docs' name `ConnectionType` for
something related but different, which risks being quite confusing.

Also move the type from the global `types.js` to the one place where
we refer to it, so there isn't such a general name floating around in
basically our global namespace; and add a TODO to move it to a still
more appropriate spot, namely a libdef.