tests: Add `eg.reduxState`, to help test selectors.
This builds on 415f0546b, with the same purpose of helping us to write
unit tests that
 * pass genuine well-typed data,
 * and don't have a lot of repetitive boilerplate for the boring bits.

We have a lot of tests for selectors, and they generally need a value
of type `GlobalState` to pass to the selector.  This lets us start
doing that.  Also demonstrate it by converting a couple of test cases
to the nice new form.