subscriptions: Delete subscription when a stream is deleted.
Fixes #3474.

When a stream is deleted, any subscription to it also must cease
to exist.  We'd implicitly assumed the server would send a separate
event to remove the subscription, but it doesn't; instead, that's
just part of the semantics of the stream-delete event.  So, act on
the stream-delete event to remove any subscription as well.

[greg: revised expectations of server behavior in commit message.]