msglist: Add slide-in timestamp labels for each message.
When we show several consecutive messages from the same sender
without an intervening recipient header or date divider, we don't
repeat the sender.  This has also meant that there's
 * no way to tell where one message ends and another begins;
 * and no way to see the time the later messages were sent.

This commit aims to solve both with a sliding label which shows the
timestamp of the message if clicked on it.  We do this by replacing
the existing timestamp with another component in the message list
called `timestampHtml`.  Timestamp starts out hidden on
sender-implicit messages and 'expanded' on sender-showing messages.

Browsers, by default, increase the size of the viewport to fit the
size of any element inside it.  This causes the horizontal size of
the viewport to increase if we attempt to 'slide' an element out of
the viewport.  The only way to not allow the browser to increase the
size is if there were an element which is always inside the browser,
and another element (our actual timestamp) nested inside it which
moves around inside our outer element (which has the overflow
property set to hidden).  Doing this causes the browser to think that
we're not moving the element out of the viewport, but that it's
simply overflowing from the outer element.

Fixes #3375.