compose-box: Keep input to one line when crowded by menu.
When the `ComposeBox` is expanded to show the `ComposeMenu`, viewing
streams with long names and/or with phones on small screens, the
hint text on the input gets compressed, and since it is multiline,
expands into multiple lines. This is not desirable, since there is a
chance that the text in the input box (when expanded) gets to only
one letter thick, and that would cause the area of the `ComposeBox`
to be very big.

This commit fixes the issue.