submessage: Match style and simplify in `messages` reducer.
The use of `message.id` made it a bit more complicated to understand
what this code does and be convinced it works right than it needs to
be: we get the message `message` found at key `action.message_id`, but
then our updated message gets stored back not at `action.message_id`
but instead at `message.id`.  That means this code didn't have the
shape of a straightforward "update the subobject found at some path".

The behavior ought to be the same given the `Message` type and the
invariants of this `messages` reducer as a whole... but instead, match
the other branches of this reducer and just do the dead-simple thing
where you can look 100% locally and confirm that the exact same key is
used both times.